Implement role based access control and jwt authentication
Add role-based authorization and support worker management; refine booking flows and DB mappings.

Key changes:
- Added SupportWorker model, DB mapping and SupportWorkersController with validation, unique-email checks and Coordinator-only endpoints.
- Updated BookingsController to enforce authorization, extract user/role claims, restrict actions by role, and return clearer error responses.
- Modified IBookingService and BookingService signatures/logic to accept role and userId, enforce ownership rules, validate status transitions, and use byte for booking status; removed UserId from BookingCreateDto (user is taken from claims).
- Extended application_db_context with SupportWorkers DbSet and explicit table/column mappings for services and bookings.
- Adjusted Booking model column types/names and formatting.
- Hardened AuthService: trims emails, requires @gmail.com addresses, normalizes/validates roles, and returns clearer messages.
- Program/Startup tweaks: load JWT settings directly from configuration, set RoleClaimType for JWT auth, adjust Swagger doc info, register services, and reorder middleware (error handling before auth/authorization).

These changes centralize security via JWT claims, add role restrictions (Participant/Coordinator), improve validation and error handling, and expose a new support-workers API for coordinators.

+++ NDISPortal.API/Models/Booking.cs
@@ -1,39 +1,39 @@
-﻿using Register.API.Models;
-using Service.API.Model;
-using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
+﻿    using Register.API.Models;
+    using Service.API.Model;
+    using System.ComponentModel.DataAnnotations;
+    using System.ComponentModel.DataAnnotations.Schema;
 
-namespace NdisPortal.BookingsApi.Models;
+    namespace NdisPortal.BookingsApi.Models;
 
-[Table("bookings")]
-public class Booking
-{
-    [Key]
-    [Column("id")]
-    public int Id { get; set; }
+    [Table("bookings")]
+    public class Booking
+    {
+        [Key]
+        [Column("id")]
+        public int Id { get; set; }
 
-    [Column("user_id")]
-    public int UserId { get; set; }
+        [Column("user_id")]
+        public int UserId { get; set; }
 
-    public User User { get; set; }
+        public User User { get; set; }
 
-    [Column("service_id")]
-    public int ServiceId { get; set; }
+        [Column("service_id")]
+        public int ServiceId { get; set; }
 
-    public ServiceItem Service { get; set; }
+        public ServiceItem Service { get; set; }
 
-    [Column("booking_date")]
-    public DateTime BookingDate { get; set; }
+        [Column("booking_date")]
+        public DateTime BookingDate { get; set; }
 
-    [Column("notes")]
-    public string? Notes { get; set; }
+        [Column("notes")]
+        public string? Notes { get; set; }
 
-    [Column("status")]
-    public int Status { get; set; }
+        [Column("status")]
+        public byte Status { get; set; }
 
-    [Column("created_date")]
-    public DateTime CreatedDate { get; set; }
+        [Column("created_date")]
+        public DateTime CreatedDate { get; set; }
 
-    [Column("modified_date")]
-    public DateTime ModifiedDate { get; set; }
-}
+        [Column("modified_date")]
+        public DateTime ModifiedDate { get; set; }
+    }
